Make the -s option imply the -nc options
This makes "secure" passwords more secure, without breaking backwards
compatibility with older pwgen programs.  The exact definition of
"more secure" was never defined in the man pages, so this gives us
some more wiggle room.

Update the pwgen man page to make it clear that the future behavior of
pwgen when stdout is not a tty may change further.

Addresses-Debian-Bug: #427969

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>