Teach pwgen the --remove-chars option
The --remove-chars (or -r) option removes characters from the set of
characters used to generate passwords.

This feature was suggested in Florian Lohoff in the Debian Bug
arguments to remove that would cause the original patch to hang in an
infinite loop.  The algorithm used is also more efficient than the
original patch.

Addresses-Debian-Bug: #836334

Signed-off-by: Theodore Ts'o <tytso@mit.edu>